Run the change coordinator command before excluding the coordinator
