+++ controllers/change_coordinators.go
@@ -22,10 +22,7 @@ package controllers
 
 import (
 	"context"
-	"fmt"
-	"math"
-	"strings"
-
+	"github.com/FoundationDB/fdb-kubernetes-operator/internal/coordinator"
 	"github.com/FoundationDB/fdb-kubernetes-operator/internal/locality"
 	"github.com/go-logr/logr"
 
@@ -59,8 +56,8 @@ func (c changeCoordinators) reconcile(ctx context.Context, r *FoundationDBCluste
 	}
 
 	coordinatorStatus := make(map[string]bool, len(status.Client.Coordinators.Coordinators))
-	for _, coordinator := range status.Client.Coordinators.Coordinators {
-		coordinatorStatus[coordinator.Address.String()] = false
+	for _, coord := range status.Client.Coordinators.Coordinators {
+		coordinatorStatus[coord.Address.String()] = false
 	}
 
 	hasValidCoordinators, allAddressesValid, err := locality.CheckCoordinatorValidity(logger, cluster, status, coordinatorStatus)
@@ -93,139 +90,15 @@ func (c changeCoordinators) reconcile(ctx context.Context, r *FoundationDBCluste
 	logger.Info("Changing coordinators")
 	r.Recorder.Event(cluster, corev1.EventTypeNormal, "ChangingCoordinators", "Choosing new coordinators")
 
-	coordinators, err := selectCoordinators(logger, cluster, status)
+	err = coordinator.ChangeCoordinators(logger, adminClient, cluster, status)
 	if err != nil {
 		return &requeue{curError: err, delayedRequeue: true}
 	}
 
-	coordinatorAddresses := make([]fdbv1beta2.ProcessAddress, len(coordinators))
-	for index, process := range coordinators {
-		coordinatorAddresses[index] = getCoordinatorAddress(cluster, process)
-	}
-
-	logger.Info("Final coordinators candidates", "coordinators", coordinatorAddresses)
-	connectionString, err := adminClient.ChangeCoordinators(coordinatorAddresses)
-	if err != nil {
-		return &requeue{curError: err, delayedRequeue: true}
-	}
-	cluster.Status.ConnectionString = connectionString
 	err = r.updateOrApply(ctx, cluster)
 	if err != nil {
 		return &requeue{curError: err, delayedRequeue: true}
 	}
 
 	return nil
-}
-
-// selectCandidates is a helper for Reconcile that picks non-excluded, not-being-removed class-matching process groups.
-func selectCandidates(cluster *fdbv1beta2.FoundationDBCluster, status *fdbv1beta2.FoundationDBStatus) ([]locality.Info, error) {
-	candidates := make([]locality.Info, 0, len(status.Cluster.Processes))
-	for _, process := range status.Cluster.Processes {
-		if process.Excluded || process.UnderMaintenance {
-			continue
-		}
-
-		if !cluster.IsEligibleAsCandidate(process.ProcessClass) {
-			continue
-		}
-
-		// Ignore processes with missing locality, see: https://github.com/FoundationDB/fdb-kubernetes-operator/issues/1254
-		if len(process.Locality) == 0 {
-			continue
-		}
-
-		// If the cluster should be using DNS in the cluster file we should make sure the locality is set.
-		if cluster.UseDNSInClusterFile() {
-			_, ok := process.Locality[fdbv1beta2.FDBLocalityDNSNameKey]
-			if !ok {
-				continue
-			}
-		}
-
-		if cluster.ProcessGroupIsBeingRemoved(fdbv1beta2.ProcessGroupID(process.Locality[fdbv1beta2.FDBLocalityInstanceIDKey])) {
-			continue
-		}
-
-		currentLocality, err := locality.InfoForProcess(process, cluster.Spec.MainContainer.EnableTLS)
-		if err != nil {
-			return nil, err
-		}
-
-		priority := cluster.GetClassCandidatePriority(process.ProcessClass)
-		// If the process is not running in the desired version or the binary is running from the shared volumes
-		// that means this process is pending a Pod recreation and will therefore be down for some time.
-		// We reduce the priority in this case to reduce the risk of successive coordinator changes. Reducing the
-		// priority should help in reducing the overall coordinator changes.
-		// See: https://github.com/FoundationDB/fdb-kubernetes-operator/issues/2015
-		if process.Version != cluster.Spec.Version || strings.HasPrefix(process.CommandLine, "/var/") {
-			// math.MinInt64 is the lowest possible priority. By adding the actual priority we make sure that we
-			// still keep the priorities, even if all processes are not yet upgraded.
-			if priority < 0 {
-				priority = math.MinInt
-			} else {
-				priority += math.MinInt
-			}
-		}
-
-		currentLocality.Priority = priority
-		candidates = append(candidates, currentLocality)
-	}
-
-	return candidates, nil
-}
-
-func selectCoordinators(logger logr.Logger, cluster *fdbv1beta2.FoundationDBCluster, status *fdbv1beta2.FoundationDBStatus) ([]locality.Info, error) {
-	var err error
-	coordinatorCount := cluster.DesiredCoordinatorCount()
-
-	candidates, err := selectCandidates(cluster, status)
-	if err != nil {
-		return nil, err
-	}
-
-	coordinators, err := locality.ChooseDistributedProcesses(cluster, candidates, coordinatorCount, locality.ProcessSelectionConstraint{
-		HardLimits:            locality.GetHardLimits(cluster),
-		SelectingCoordinators: true,
-	})
-
-	logger.Info("Current coordinators", "coordinators", coordinators, "error", err)
-	if err != nil {
-		return nil, err
-	}
-
-	coordinatorStatus := make(map[string]bool, len(status.Client.Coordinators.Coordinators))
-	for _, coordinator := range coordinators {
-		coordinatorStatus[getCoordinatorAddress(cluster, coordinator).String()] = false
-	}
-
-	hasValidCoordinators, allAddressesValid, err := locality.CheckCoordinatorValidity(logger, cluster, status, coordinatorStatus)
-	if err != nil {
-		return nil, err
-	}
-
-	if !hasValidCoordinators {
-		return nil, fmt.Errorf("new coordinators are not valid")
-	}
-
-	if !allAddressesValid {
-		return nil, fmt.Errorf("new coordinators contain invalid addresses")
-	}
-
-	return coordinators, nil
-}
-
-func getCoordinatorAddress(cluster *fdbv1beta2.FoundationDBCluster, locality locality.Info) fdbv1beta2.ProcessAddress {
-	dnsName := locality.LocalityData[fdbv1beta2.FDBLocalityDNSNameKey]
-
-	address := locality.Address
-
-	if cluster.UseDNSInClusterFile() && dnsName != "" {
-		return fdbv1beta2.ProcessAddress{
-			StringAddress: dnsName,
-			Port:          address.Port,
-			Flags:         address.Flags,
-		}
-	}
-
-	return address
 }
